requirements.txt
absl-py==2.3.1
accelerate==1.9.0
addict==2.4.0
aiofiles==23.2.1
aiosignal==1.4.0
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
autograd==1.8.0
black==21.4b2
blis==0.7.11
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clip==0.2.0
cloudpathlib==0.16.0
cloudpickle==3.1.1
cma==2.7.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
decorator==4.4.2
decord==0.6.0
deepspeed==0.12.3
detectron2 @ git+https://github.com/MaureenZOU/detectron2-xyz.git@42121d75e10d9f858f3a91b6a39f5722c02868f0
editdistance==0.6.2
einops==0.8.0
einx==0.3.0
encodec==0.1.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fairscale==0.4.0
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.58.5
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20221221
GPy==1.13.2
gradio==4.26.0
gradio_client==0.15.1
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
hydra-core==1.3.2
idna==3.10
imagecodecs==2025.3.30
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
iopath==0.1.9
ipdb==0.13.13
ipython==8.37.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.1
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
Levenshtein==0.27.1
librosa==0.9.0
llvmlite==0.44.0
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mmcls==0.25.0
mmcv==1.4.2
modelscope_studio @ http://thunlp.oss-cn-qingdao.aliyuncs.com/multi_modal/never_delete/modelscope_studio-0.4.0.9-py3-none-any.whl#sha256=ff8745a9580789d9b9edae9a4101192bbd300aba91cb8779187f1a2e0d67028f
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.13
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.11.1.1
nltk==3.8.1
numba==0.61.2
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open_clip_torch==3.0.0
opencv-python==4.11.0.86
opencv-python-headless==4.5.5.64
openpyxl==3.1.2
orjson==3.10.18
packaging==23.2
pandas==2.3.1
paramz==0.9.6
parso==0.8.4
pathspec==0.12.1
peft==0.14.0
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.3.8
pooch==1.8.2
portalocker==3.2.0
preshed==3.0.10
proglog==0.1.12
prompt_toolkit==3.0.51
protobuf==4.25.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==4.0.1
pydub==0.25.1
pyfiglet==1.0.3
Pygments==2.19.2
pymoo==0.5.0
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
resampy==0.4.3
rich==13.9.4
rpds-py==0.26.0
ruff==0.12.3
sacrebleu==2.3.2
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.12.0
seaborn==0.13.0
semantic-version==2.10.0
sentence-transformers==5.0.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.11
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
socksio==1.0.0
soundfile==0.12.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
thinc==8.2.5
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.19
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
toolz==1.0.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.66.1
traitlets==5.14.3
transformers==4.44.2
triton==2.2.0
typer==0.9.4
typing_extensions==4.8.0
tzdata==2025.2
ultralytics==8.3.178
ultralytics-thop==2.0.15
urllib3==2.5.0
uvicorn==0.24.0.post1
vector-quantize-pytorch==1.18.5
vocos==0.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.1.3
yacs==0.1.8
yapf==0.43.0
